+++ tests/testthat/test-fit.R
@@ -439,6 +439,100 @@ test_that("mmrm works for constructed control", {
   ))
 })
 
+test_that("mmrm works for start = NULL", {
+  expect_silent(mmrm(
+    FEV1 ~ ARMCD + ar1(AVISIT | SEX / USUBJID),
+    data = fev_data,
+    reml = TRUE,
+    control = mmrm_control(optimizer = c("BFGS", "CG"), start = NULL)
+  ))
+})
+
+## start values ----
+
+test_that("start can be emp_start or std_start and they work", {
+  expect_silent(
+    mmrm(
+      formula = FEV1 ~ ARMCD + us(AVISIT | USUBJID),
+      data = fev_data,
+      start = emp_start
+    )
+  )
+  expect_silent(
+    mmrm(
+      formula = FEV1 ~ ARMCD + us(AVISIT | USUBJID),
+      data = fev_data,
+      start = std_start
+    )
+  )
+})
+
+test_that("mmrm fails if start function does not provide correct values", {
+  wrong_start <- function(...) {
+    c(NA_real_, 0L)
+  }
+  expect_error(
+    mmrm(
+      formula = FEV1 ~ ARMCD + us(AVISIT | USUBJID),
+      data = fev_data,
+      start = wrong_start
+    ),
+    "Assertion on 'start_values' failed: Must have length 10, but has length 2."
+  )
+  expect_error(
+    mmrm(
+      formula = FEV1 ~ ARMCD + ar1(AVISIT | USUBJID),
+      data = fev_data,
+      start = wrong_start
+    ),
+    "Assertion on 'start_values' failed: Contains missing values \\(element 1\\)"
+  )
+})
+
+test_that("mmrm fails if start function fails", {
+  wrong_start <- function(...) {
+    stop("this is wrong")
+  }
+  expect_error(
+    mmrm(
+      formula = FEV1 ~ ARMCD + us(AVISIT | USUBJID),
+      data = fev_data,
+      start = wrong_start
+    ),
+    "this is wrong"
+  )
+})
+
+test_that("mmrm works for transformed response using emp_start", {
+  expect_silent(
+    mmrm(
+      formula = log(FEV1) ~ ARMCD + exp(FEV1_BL) + us(AVISIT | USUBJID),
+      data = fev_data,
+      start = std_start
+    )
+  )
+  expect_silent(
+    mmrm(
+      formula = log(FEV1) ~ ARMCD + exp(FEV1_BL) + us(AVISIT | USUBJID),
+      data = fev_data,
+      start = emp_start
+    )
+  )
+})
+
+test_that("mmrm works for character covariate using emp_start", {
+  fev_data2 <- fev_data
+  fev_data2$ARMCD <- as.character(fev_data2$ARMCD)
+  fev_data2$USUBJID <- as.character(fev_data2$USUBJID)
+  expect_silent(
+    mmrm(
+      formula = log(FEV1) ~ ARMCD + exp(FEV1_BL) + us(AVISIT | USUBJID),
+      data = fev_data2,
+      start = emp_start
+    )
+  )
+})
+
 test_that("mmrm still works for deprecated 'automatic' optimizer", {
   expect_warning(
     mmrm(